+++ packages/react-charting/src/components/AreaChart/AreaChartRTL.test.tsx
@@ -283,6 +283,8 @@ const chartDataWithDates = {
   pointLineOptions: { strokeWidth: 2, strokeDasharray: '10 10', stroke: DefaultPalette.blueDark },
 };
 
+const secondaryYScalePoints = [{ yMaxValue: 50000, yMinValue: 10000 }];
+
 describe('Area chart rendering', () => {
   beforeEach(updateChartWidthAndHeight);
   afterEach(sharedAfterEach);
@@ -337,6 +339,17 @@ describe('Area chart rendering', () => {
       !(isTimezoneSet(Timezone.UTC) && isTestEnv()),
     );
   });
+
+  testWithoutWait(
+    'Should render the Area chart with secondary Y axis',
+    AreaChart,
+    { data: chartData, secondaryYScaleOptions: secondaryYScalePoints },
+    container => {
+      // Assert
+      expect(getById(container, /yAxisGElementSecondarychart_/i)).toBeDefined();
+      expect(container).toMatchSnapshot();
+    },
+  );
 });
 
 describe('Area chart - Subcomponent Area', () => {