+++ packages/react/src/components/MessageBar/MessageBar.types.ts
@@ -128,6 +128,13 @@ export interface IMessageBarProps extends React.HTMLAttributes<HTMLElement>, Rea
    * @default true
    */
   delayedRender?: boolean;
+
+  /**
+   * An optional override to show the expand/collapse icon. It will only be shown by default for
+   * single-line truncated MessageBars that do not have actions.
+   * @defaultvalue false
+   */
+  showExpandButton?: boolean;
 }
 
 /**